requirements.txt
# Core dependencies
pyserial>=3.5  # For UART radio programming

# Optional dependencies for development:
# pytest>=7.0    # For testing
# pytest-cov>=4.0  # For coverage reports
